refactor: change structure of component BookContent
