--- src/widgets/BookContent/ui/WordsProgress.tsx
@@ -1,13 +1,0 @@
-import { Stack, CircularProgress } from '@mui/material';
-
-const styleProgress = { height: '25vw', justifyContent: 'center', alignItems: 'center' };
-
-const WordsProgress = () => {
-  return (
-    <Stack sx={styleProgress}>
-      <CircularProgress color="info" size={60} />
-    </Stack>
-  );
-};
-
-export default WordsProgress;
